Move IdApiJsonBodyParser
