+++ identity/app/idapiclient/parser/IdApiJsonBodyParser.scala
@@ -1,9 +1,8 @@
-package idapiclient
+package idapiclient.parser
 
-import net.liftweb.json.JsonAST.JValue
 import com.gu.identity.model.{LiftJsonConfig, Error => IdApiError}
-import idapiclient.parser.{JodaJsonSerializer, JsonBodyParser}
 import idapiclient.responses.Error
+import net.liftweb.json.JsonAST.JValue
 import utils.SafeLogging
 
 class IdApiJsonBodyParser extends JsonBodyParser with SafeLogging {
